fix some wtf segfaults by calling gc.collect()
When closing a thread view window, python might free the "Threads"
object before its "Thread" children, which is wrong because notmuch
expects the reverse order (and that's why notmuch keeps "_parent"
pointers).

The reason is unknown. Also, it doesn't happen when just closing the
threads list, only with the thread view.

The problem seems to be fixed when calling "gc.collect()" manually when
closing the thread view.