introduce "trashing" filter plugin
This filter plugin will search messages with tag "deleted" and move them
to a "Trash" folder, and move to inbox messages that are in "Trash"
folder but don't have "deleted" tag. It also operates as soon as tag
"deleted" is set or unset (it needs to be done through UI to be detected
as soon as possible).

Later, it may be able to use maildir flag "trashed" instead of moving to
"Trash" folder.