ui.models.ThreadListModel: emit dataChanged a row when tags change
When a tag is added to a displayed thread's messages, the view should be
updated to reflect that change.

Note a thread will not appear if it suddenly matches the selected query
string, and a thread will not be hidden if it doesn't match the query
string anymore (for example, all its messages are tagged "deleted").
Only re-running the search can take care of those.