config: add a "credentials" system that saves secrets in memory
This credentials system if optional if not desired.

The use case is when gpg is used (via "pass" for example), the keyring
password may be asked too often, everytime mails are fetched or sent,
which can be painful. The credentials system solves it by still using
gpg for securely storing passwords offline, but keeping them decrypted
in memory so they can be fed to plugins without needing user unlock.

From a security POV, this means passwords could be read in the process
memory anytime, while without "credentials", they could be read only
when the user had to decrypt the password, i.e when mail is fetched or
sent, which isn't that uncommon in a mail app, so the overall security
should not really be lowered.