setup: .ui files weren't included as package data and some forgotten
"include_package_data" was not in the right setup.cfg section, so *.ui
files were not placed along *.py files so ui might not be loaded
correctly (if not compiled to *_ui.py files).

Also, builtin plugins can have their own *.ui files, so include them in
MANIFEST.in too.